Refactor code so that AttestationGenerator is now in state transition test support
