+++ ethereum/datastructures/src/main/java/tech/pegasys/artemis/datastructures/util/AttestationUtil.java
@@ -15,6 +15,9 @@ package tech.pegasys.artemis.datastructures.util;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static java.lang.Math.toIntExact;
+import static tech.pegasys.artemis.datastructures.util.BeaconStateUtil.compute_start_slot_at_epoch;
+import static tech.pegasys.artemis.datastructures.util.BeaconStateUtil.get_block_root_at_slot;
+import static tech.pegasys.artemis.datastructures.util.BeaconStateUtil.get_current_epoch;
 import static tech.pegasys.artemis.datastructures.util.BeaconStateUtil.get_domain;
 import static tech.pegasys.artemis.datastructures.util.CommitteeUtil.get_beacon_committee;
 import static tech.pegasys.artemis.util.bls.BLSAggregate.bls_aggregate_pubkeys;
@@ -31,11 +34,13 @@ import java.util.stream.Collectors;
 import org.apache.logging.log4j.Level;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.artemis.datastructures.blocks.BeaconBlock;
 import tech.pegasys.artemis.datastructures.operations.Attestation;
 import tech.pegasys.artemis.datastructures.operations.AttestationData;
 import tech.pegasys.artemis.datastructures.operations.AttestationDataAndCustodyBit;
 import tech.pegasys.artemis.datastructures.operations.IndexedAttestation;
 import tech.pegasys.artemis.datastructures.state.BeaconState;
+import tech.pegasys.artemis.datastructures.state.Checkpoint;
 import tech.pegasys.artemis.datastructures.state.Validator;
 import tech.pegasys.artemis.util.SSZTypes.Bitlist;
 import tech.pegasys.artemis.util.SSZTypes.SSZList;
@@ -293,5 +298,22 @@ public class AttestationUtil {
     } else {
       return false;
     }
+  }
+
+  // Get attestation data that does not include attester specific shard or crosslink information
+  public static AttestationData getGenericAttestationData(BeaconState state, BeaconBlock block) {
+    UnsignedLong slot = state.getSlot();
+    // Get variables necessary that can be shared among Attestations of all validators
+    Bytes32 beacon_block_root = block.signing_root("signature");
+    UnsignedLong start_slot = compute_start_slot_at_epoch(get_current_epoch(state));
+    Bytes32 epoch_boundary_block_root =
+        start_slot.compareTo(slot) == 0
+            ? block.signing_root("signature")
+            : get_block_root_at_slot(state, start_slot);
+    Checkpoint source = state.getCurrent_justified_checkpoint();
+    Checkpoint target = new Checkpoint(get_current_epoch(state), epoch_boundary_block_root);
+
+    // Set attestation data
+    return new AttestationData(slot, UnsignedLong.ZERO, beacon_block_root, source, target);
   }
 }
