+++ ethereum/statetransition/src/main/java/tech/pegasys/artemis/statetransition/CommitteeAssignment.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.validator.client;
+package tech.pegasys.artemis.statetransition;
 
 import com.google.common.primitives.UnsignedLong;
 import java.util.List;
